% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{molit.rt.apt}
\alias{molit.rt.apt}
\title{아파트 실거래가}
\format{Classes ‘data.table’ and 'data.frame'}
\usage{
molit.rt.apt
}
\description{
이 데이터는 http://rt.molit.go.kr/에서 제공하는 아파트 매매 및 전세 실거래가 정보입니다.\cr
현재 2016년 1월부터 8월까지의 데이터가 생성되어있습니다.\cr
면적은 m^2, 금액은 만원단위로 표시됩니다.\cr
}
\author{
Jaehyun Shin \email{keepcosmos@gmail.com}
}
\references{
\url{http://rtdown.molit.go.kr}
}
\keyword{datasets}

