% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\encoding{UTF-8}
\name{molit.rt.get}
\alias{molit.rt.get}
\title{molit.rt.get}
\usage{
molit.rt.get(type = NA, region = NA, colnames.locale = "en")
}
\arguments{
\item{type}{House Type (sh, rh, apt)}

\item{region}{Location name}

\item{engColNames}{Convert to English column names, default TRUE}
}
\description{
Get data by type and region name
}
\examples{
    molit.rt.get("sh", address)
    molit.rt.get("apt", address)
}

