% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molit.rt.sh}
\alias{molit.rt.sh}
\alias{molit.rt.sh.2016}
\title{molit.rt.sh}
\format{Classes 'data.table' and 'data.frame'}
\usage{
molit.rt.sh.2016
}
\description{
Single-Family houses Real Trading & Rent Datasets from 2006
source from molit.go.kr\cr
}
\details{
Call 'molit.rt.sh.(year)'
}
\author{
Jaehyun Shin \email{keepcosmos@gmail.com}
}
\references{
\url{http://rtdown.molit.go.kr}
}
\keyword{datasets}

