% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molit.rt.apt}
\alias{molit.rt.apt}
\alias{molit.rt.apt.2016}
\title{molit.rt.apt}
\format{Classes 'data.table' and 'data.frame'}
\usage{
molit.rt.apt.2016
}
\description{
Apartments Real Trading & Rent Datasets from 2006\cr
source from molit.go.kr\cr
\cr
Call 'molit.rt.apt.(year)'
}
\author{
Jaehyun Shin \email{keepcosmos@gmail.com}
}
\references{
\url{http://rtdown.molit.go.kr}
}
\keyword{datasets}

