% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molit.rt.rh}
\alias{molit.rt.rh}
\alias{molit.rt.rh.2016}
\title{molit.rt.rh}
\format{Classes 'data.table' and 'data.frame'}
\usage{
molit.rt.rh.2016
}
\description{
Row houses(tenement house) Real Trading & Rent Datasets from 2006\cr
source from molit.go.kr\cr
}
\details{
#' Call 'molit.rt.rh.(year)'
}
\author{
Jaehyun Shin \email{keepcosmos@gmail.com}
}
\references{
\url{http://rtdown.molit.go.kr}
}
\keyword{datasets}

